requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Machine Learning
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
